+++ Backend/demo/src/main/java/com/example/demo/hooks/NotificationDTO.java
@@ -4,18 +4,18 @@ import java.time.LocalDateTime;
 
 public class NotificationDTO {
 
-    private Long id;
+    private int id;
     private LocalDateTime dateOfSend;
     private String content;
     private Boolean isRead;
-    private Long supporterId; 
+    private int supporterId; 
 
     
     public NotificationDTO() {}
 
     
-    public NotificationDTO(Long id, LocalDateTime dateOfSend, String content,
-                           Boolean isRead, Long supporterId) {
+    public NotificationDTO(int id, LocalDateTime dateOfSend, String content,
+                           Boolean isRead, int supporterId) {
         this.id = id;
         this.dateOfSend = dateOfSend;
         this.content = content;
@@ -24,8 +24,8 @@ public class NotificationDTO {
     }
 
     
-    public Long getId() { return id; }
-    public void setId(Long id) { this.id = id; }
+    public int getId() { return id; }
+    public void setId(int id) { this.id = id; }
 
     public LocalDateTime getDateOfSend() { return dateOfSend; }
     public void setDateOfSend(LocalDateTime dateOfSend) { this.dateOfSend = dateOfSend; }
@@ -36,6 +36,6 @@ public class NotificationDTO {
     public Boolean getIsRead() { return isRead; }
     public void setIsRead(Boolean isRead) { this.isRead = isRead; }
 
-    public Long getSupporterId() { return supporterId; }
-    public void setSupporterId(Long supporterId) { this.supporterId = supporterId; }
+    public int getSupporterId() { return supporterId; }
+    public void setSupporterId(int supporterId) { this.supporterId = supporterId; }
 }
